enrichment/requirements.txt
# Enrichment Service Requirements
# Enterprise Banking SOC - Threat Intelligence and LLM Integration

# Core frameworks
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.1.1
httpx==0.24.1

# LLM Libraries
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
langchain-community==0.0.1
transformers==4.35.0
torch==2.0.1
sentence-transformers==2.2.2
llama-cpp-python==0.2.20

# CVE and Vulnerability Data
nvdlib==0.7.4
cvelib==1.1.0
pyexploitdb==0.2.3

# Threat Intelligence
stix2==3.0.1
taxii2-client==2.3.0
pymisp==2.4.170
pycti==5.12.17
yara-python==4.3.1

# MITRE ATT&CK
mitreattack-python==3.0.4
attackcti==0.3.4

# Data processing
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.0
beautifulsoup4==4.12.2

# Database and storage
opensearch-py==2.3.1
redis==4.6.0
neo4j==5.14.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
jinja2==3.1.2
jsonschema==4.19.0

# Logging and monitoring
structlog==23.1.0
prometheus-client==0.17.1
sentry-sdk==1.29.2

# Security
cryptography==41.0.3
jwt==1.3.1

# Caching and performance
aiocache==0.12.2
cachetools==5.3.1

# Asset management
cmdbuild==0.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-mock==3.11.1